Avoid gcc warnings.

+++ tests/test-crc.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2005, 2006 Free Software Foundation
+ * Copyright (C) 2005, 2006, 2007 Free Software Foundation
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -34,28 +34,28 @@ main (int argc, char *argv[])
   p = crc32_update_no_xor (42, "foo", 3);
   if (p != 0x46e87f05)
     {
-      printf ("cunx got %lx\n", p);
+      printf ("cunx got %lx\n", (unsigned long) p);
       return 1;
     }
 
   p = crc32_no_xor ("foo", 3);
   if (p != 0x7332bc33)
     {
-      printf ("cnx got %lx\n", p);
+      printf ("cnx got %lx\n", (unsigned long) p);
       return 1;
     }
 
   p = crc32_update (42, "foo", 3);
   if (p != 0xb9a9a617)
     {
-      printf ("cu got %lx\n", p);
+      printf ("cu got %lx\n", (unsigned long) p);
       return 1;
     }
 
   p = crc32 ("foo", 3);
   if (p != 0x8c736521)
     {
-      printf ("c got %lx\n", p);
+      printf ("c got %lx\n", (unsigned long) p);
       return 1;
     }
